deployments/mantarray/services/db-viewer/src/requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
fastapi==0.70.1
h11==0.12.0
idna==3.3
pydantic==1.9.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
uvicorn==0.16.0
<<<<<<< HEAD

boto3==1.18.5
jinja2==3.0.3
psycopg2-binary==2.9.3
immutabledict==2.2.1
=======
PyMySQL==1.0.2
jinja2==3.0.3
>>>>>>> renamed deployment
